Split out the conversation options to a module so it can be used with the runner
