+++ spec/converse/conversation_spec.rb
@@ -13,37 +13,6 @@ RSpec.describe Converse::Conversation do
     it "sets the current step to the template" do
       expect(subject.steps.first).to eq proc
     end
-
-    context "valid options" do
-      it "allows for empty options" do
-        expect { described_class.new(template, {}) }.to_not raise_error
-      end
-
-      it "allows for an access token" do
-        expect { described_class.new(template, { access_token: "BLAH" }) }.to_not raise_error
-      end
-
-      it "allows for an author id" do
-        expect { described_class.new(template, { author_id: "BLAH" }) }.to_not raise_error
-      end
-
-      it "allows the options to respond as methods on the conversation" do
-        subject = described_class.new(template, { author_id: "BLAH" })
-
-        expect(subject.author_id).to eq "BLAH"
-      end
-
-      it "does not respond to messages that are not options" do
-        expect { subject.blah }.to raise_error NoMethodError
-      end
-    end
-
-    context "invalid options" do
-      it "does not allow for random options" do
-        expect { described_class.new(template, { blah: :bleh }) }.to \
-          raise_error Converse::InvalidOptionsError, "blah is not a valid option"
-      end
-    end
   end
 
   describe "#ask" do
